requirements.txt
pyrogram==2.0.106
python-dotenv==1.0.1
telegraph==2.2.0
tgcrypto==1.2.5
asyncio==3.4.3
uuid==1.30
requests
psycopg2
pyyaml